refactor: fsd에 맞게 수정
