+++ src/features/comment-management/api/comment-mutations.ts
@@ -1,5 +1,100 @@
+// import { useMutation, useQueryClient } from '@tanstack/react-query';
+// import { addComment, deleteComment, likeComment, updateComment } from '../../../shared/api/comments';
+// import { Comment, CommentsResponse, EditableComment, NewComment } from '../../../entities/comment/model/types';
+
+// export const useAddCommentMutation = () => {
+//   const queryClient = useQueryClient();
+
+//   return useMutation({
+//     mutationFn: (newComment: NewComment) => addComment(newComment),
+//     onSuccess: (comment) => {
+//       queryClient.setQueryData(['comments', comment.postId], (oldData: CommentsResponse) => {
+//         if (!oldData) return { comments: [comment] };
+
+//         return {
+//           ...oldData,
+//           comments: [...oldData.comments, comment],
+//         };
+//       });
+//     },
+//   });
+// };
+
+
+// export const useUpdateCommentMutation = () => {
+//   const queryClient = useQueryClient();
+
+//   return useMutation({
+//     mutationFn: (comment: EditableComment) => updateComment(comment),
+//     onSuccess: (updatedComment) => {
+//       queryClient.setQueryData(['comments', updatedComment.postId], (oldData: CommentsResponse) => {
+//         if (!oldData) return oldData;
+
+//         return {
+//           ...oldData,
+//           comments: oldData.comments.map((comment: Comment) =>
+//             comment.id === updatedComment.id ? { ...comment, ...updatedComment } : comment
+//           ),
+//         };
+//       });
+//     },
+//   });
+// };
+
+// export const useDeleteCommentMutation = () => {
+//   const queryClient = useQueryClient();
+
+//   return useMutation({
+//     mutationFn: ({ id }: { id: number; postId: number }) => deleteComment(id),
+//     onSuccess: (_, variables) => {
+//       const { id, postId } = variables;
+
+//       queryClient.setQueryData(['comments', postId], (oldData: CommentsResponse) => {
+//         if (!oldData) return oldData;
+
+//         return {
+//           ...oldData,
+//           comments: oldData.comments.filter((comment: Comment) => comment.id !== id),
+//         };
+//       });
+
+//     },
+//   });
+// };
+
+
+// export const useLikeCommentMutation = () => {
+//   const queryClient = useQueryClient();
+
+//   return useMutation({
+//     mutationFn: ({ id, likes }: { id: number; likes: number; postId: number }) =>
+//       likeComment(id, likes),
+//     onSuccess: (updatedComment, variables) => {
+//       const postId = variables.postId;
+//       const commentId = variables.id;
+//       const currentLikes = variables.likes;
+
+//       queryClient.setQueryData(['comments', postId], (oldData: CommentsResponse) => {
+//         if (!oldData) return { comments: [updatedComment] };
+//         return {
+//           ...oldData,
+//           comments: oldData.comments.map((comment: Comment) => {
+//             if (comment.id === commentId) {
+//               return {
+//                 ...comment,
+//                 likes: currentLikes + 1
+//               };
+//             }
+//             return comment;
+//           })
+//         };
+//       });
+//     },
+//   });
+// };
+
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { addComment, deleteComment, likeComment, updateComment } from '../../../shared/api/comments';
+import { addComment, deleteComment, likeComment, updateComment } from '../../../entities/comment/api/comments';
 import { Comment, CommentsResponse, EditableComment, NewComment } from '../../../entities/comment/model/types';
 
 export const useAddCommentMutation = () => {
@@ -57,7 +152,6 @@ export const useDeleteCommentMutation = () => {
           comments: oldData.comments.filter((comment: Comment) => comment.id !== id),
         };
       });
-
     },
   });
 };
