+++ src/features/post-management/api/post-mutations.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { addPost, deletePost, updatePost } from '../../../shared/api/posts';
 import { EditablePost, NewPost, Post, PostsResponse } from '../../../entities/post/model';
+// import { addPost, deletePost, updatePost } from '../../../shared/api/posts';
+import { addPost, deletePost, updatePost } from '../../../entities/post/api/posts';
 
 export const useAddPostMutation = () => {
   const queryClient = useQueryClient();
