+++ src/features/user-management/ui/UserDetailDialog.tsx
@@ -1,11 +1,71 @@
+// import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
+// import { useModalStore } from "../../ui/model/model-store"
+// import { useUserQuery } from "../api/user-queries"
+
+// export const UserDetailDialog = () => {
+//   const { isUserDetailDialogOpen, closeUserDetailDialog, selectedUser } = useModalStore()
+
+//   const { data: userData, isLoading } = useUserQuery(selectedUser?.id || 0)
+//   if (!selectedUser) return null
+//   const user = userData || selectedUser
+
+//   return (
+//     <Dialog open={isUserDetailDialogOpen} onOpenChange={closeUserDetailDialog}>
+//       <DialogContent>
+//         <DialogHeader>
+//           <DialogTitle>사용자 정보</DialogTitle>
+//         </DialogHeader>
+
+//         {isLoading ? (
+//           <div className="flex justify-center p-4">로딩 중...</div>
+//         ) : (
+//           <div className="space-y-4">
+//             {/* 사용자 프로필 이미지 및 이름 */}
+//             <img src={user.image} alt={user.username} className="w-24 h-24 rounded-full mx-auto" />
+//             <h3 className="text-xl font-semibold text-center">{user?.username}</h3>
+
+//             {/* 사용자 상세 정보 */}
+//             <div className="space-y-2">
+//               <p>
+//                 <strong>이름:</strong> {user?.firstName} {user?.lastName}
+//               </p>
+
+//               <p>
+//                 <strong>나이:</strong> {user?.age}
+//               </p>
+
+//               <p>
+//                 <strong>이메일:</strong> {user?.email}
+//               </p>
+
+//               <p>
+//                 <strong>전화번호:</strong> {user?.phone}
+//               </p>
+
+//               <p>
+//                 <strong>주소:</strong> {user?.address?.address}, {user?.address?.city}, {user?.address?.state}
+//               </p>
+
+//               <p>
+//                 <strong>직장:</strong> {user?.company?.name} - {user?.company?.title}
+//               </p>
+//             </div>
+//           </div>
+//         )}
+//       </DialogContent>
+//     </Dialog>
+//   )
+// }
+
+import { useUserQuery } from "../../../entities/user/api/user-queries"
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../../shared/ui"
-import { useModalStore } from "../../ui/model/model-store"
-import { useUserQuery } from "../api/user-queries"
+import { useUserModalStore } from "../../modals"
 
 export const UserDetailDialog = () => {
-  const { isUserDetailDialogOpen, closeUserDetailDialog, selectedUser } = useModalStore()
+  const { isUserDetailDialogOpen, closeUserDetailDialog, selectedUser } = useUserModalStore()
 
   const { data: userData, isLoading } = useUserQuery(selectedUser?.id || 0)
+
   if (!selectedUser) return null
   const user = userData || selectedUser
