+++ src/shared/api/comments.ts
@@ -1,39 +1,39 @@
-import { api } from './base';
-import {
-  Comment,
-  CommentsResponse,
-  NewComment,
-  EditableComment
-} from '../../entities/comment/model/types';
+// import { api } from './base';
+// import {
+//   Comment,
+//   CommentsResponse,
+//   NewComment,
+//   EditableComment
+// } from '../../entities/comment/model/types';
 
 
-export const fetchCommentsByPostId = async (
-  postId: number
-): Promise<CommentsResponse> => {
-  return api.get<CommentsResponse>(`/comments/post/${postId}`);
-};
+// export const fetchCommentsByPostId = async (
+//   postId: number
+// ): Promise<CommentsResponse> => {
+//   return api.get<CommentsResponse>(`/comments/post/${postId}`);
+// };
 
-export const addComment = async (
-  newComment: NewComment
-): Promise<Comment> => {
-  return api.post<Comment>('/comments/add', newComment);
-};
+// export const addComment = async (
+//   newComment: NewComment
+// ): Promise<Comment> => {
+//   return api.post<Comment>('/comments/add', newComment);
+// };
 
-export const updateComment = async (
-  comment: EditableComment
-): Promise<Comment> => {
-  return api.put<Comment>(`/comments/${comment.id}`, { body: comment.body });
-};
+// export const updateComment = async (
+//   comment: EditableComment
+// ): Promise<Comment> => {
+//   return api.put<Comment>(`/comments/${comment.id}`, { body: comment.body });
+// };
 
-export const deleteComment = async (
-  id: number
-): Promise<void> => {
-  return api.delete<void>(`/comments/${id}`);
-};
+// export const deleteComment = async (
+//   id: number
+// ): Promise<void> => {
+//   return api.delete<void>(`/comments/${id}`);
+// };
 
-export const likeComment = async (
-  id: number,
-  currentLikes: number
-): Promise<Comment> => {
-  return api.patch<Comment>(`/comments/${id}`, { likes: currentLikes + 1 });
-};
+// export const likeComment = async (
+//   id: number,
+//   currentLikes: number
+// ): Promise<Comment> => {
+//   return api.patch<Comment>(`/comments/${id}`, { likes: currentLikes + 1 });
+// };
