+++ src/shared/api/index.ts
@@ -1,4 +1,4 @@
 export * from './base';
-export * as postsApi from './posts';
-export * as usersApi from './users';
+// export * as postsApi from './posts';
+// export * as usersApi from './users';
 // export * as commentsApi from './comments';
