+++ src/shared/api/users.ts
@@ -1,22 +1,22 @@
-import { api } from './base';
-import { User, UsersResponse, UserProfileOptions } from '../../entities/user/model';
+// import { api } from './base';
+// import { User, UsersResponse, UserProfileOptions } from '../../entities/user/model';
 
 
-export const fetchUsers = async (
-  options: UserProfileOptions = {}
-): Promise<UsersResponse> => {
-  const params = new URLSearchParams();
-  params.set('limit', '0');
+// export const fetchUsers = async (
+//   options: UserProfileOptions = {}
+// ): Promise<UsersResponse> => {
+//   const params = new URLSearchParams();
+//   params.set('limit', '0');
 
-  if (options.select) {
-    params.set('select', options.select);
-  }
+//   if (options.select) {
+//     params.set('select', options.select);
+//   }
 
-  return api.get<UsersResponse>(`/users?${params.toString()}`);
-};
+//   return api.get<UsersResponse>(`/users?${params.toString()}`);
+// };
 
-export const fetchUserById = async (
-  id: number
-): Promise<User> => {
-  return api.get<User>(`/users/${id}`);
-};
+// export const fetchUserById = async (
+//   id: number
+// ): Promise<User> => {
+//   return api.get<User>(`/users/${id}`);
+// };
