+++ src/shared/config/constants.ts
@@ -1,0 +1,13 @@
+export const SORT_OPTIONS = [
+  { value: "none", label: "없음" },
+  { value: "id", label: "ID" },
+  { value: "title", label: "제목" },
+  { value: "reactions", label: "반응" },
+];
+
+export const SORT_ORDER_OPTIONS = [
+  { value: "asc", label: "오름차순" },
+  { value: "desc", label: "내림차순" },
+];
+
+export const ITEMS_PER_PAGE_OPTIONS = [10, 20, 30, 50];
