+++ src/widgets/comment-section/index.ts
@@ -1,0 +1,1 @@
+export * from './ui/CommentSectionWidget';
